Extract batchScaler out of webhook-based autoscaler for testability
